pyproject.toml
[project]
name = "youtube-mp3-converter-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.128.6",
    "ngrok>=1.4.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "uvicorn>=0.39.0",
    "yt-dlp>=2025.10.14",
]
